Implemented first custom exception class. currently working on author entity and its crud operatinos

+++ src/main/java/org/libapp/libapp/repository/BookRepo.java
@@ -3,5 +3,5 @@ package org.libapp.libapp.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.libapp.libapp.entity.Book;
 
-public interface BookRepo extends JpaRepository<Book, Long> {
+public interface BookRepo extends JpaRepository<Book, Integer> {
 }
